kommander-app/requirements.txt
loguru
datetime
pytz
oci
py-cord
pyyaml